requirements.txt
python-dotenv~=1.0.0
colorama~=0.4.6
tiktoken~=0.3.3
pyyaml~=6.0
markdown~=3.4.3
beautifulsoup4~=4.12.2
pylatexenc~=2.10
PyPDF2~=2.11.1
python-docx~=0.8.11
openai